Remove [pytest] section of tox.ini
Pass py.test arguments in the commands statements instead of
using addopts in [pytest].

This is so that PyCharm's coverage tool can work. It doesn't like
pytest-cov, and if there's a [pytest] section in tox.ini that turns
it on, it gets picked up even by non-tox runs of py.test.